Rename RoomContainer to Room
