+++ src/app/components/lowerContainer/lowerContainer.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import './lowerContainerStyle.scss';
 import FormContainer from '../formContainer/formContainer.jsx';
-import Room from '../room/roomContainer.jsx';
+import Room from '../room/room.jsx';
 
 //render room component
 //render form container componenet
