--- src/app/components/room/roomContainer.jsx
@@ -1,24 +1,0 @@
-import React from 'react';
-import './roomContainerStyle.scss';
-import Plant from '../plants/plant.jsx';
-
-const plant = {
-  species: 'Monstera deliciosa',
-  watering_frequency_per_week: 1, 
-  humidity: 70, 
-  light: 5
-}
-
-export default function Room() {
-  
-  return (
-    <div id="roomContainer">
-      <Plant plant={plant}/>
-      <Plant plant={plant}/>
-      <Plant plant={plant}/>
-      <Plant plant={plant}/>
-      <Plant plant={plant}/>
-      <Plant plant={plant}/>
-    </div>
-  );
-}
