+++ src/app/components/room/roomStyle.scss
@@ -1,5 +1,5 @@
 @import '../../_variables.scss';
-#roomContainer {
+#room {
   background-color: $primaryWhite_opacity10;
   display: flex;
   flex-direction: row;
